backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1,<4.0"
Django = "^3.1.7"
djangorestframework = "^3.12.2"
django-cors-headers = "^3.7.0"
django-rest-auth = "^0.9.5"
sklearn = "^0.0"
scikit-learn = "^0.24.1"
pandas = "^1.2.3"
coverage = "^5.5"
djangorestframework-jwt = "^1.11.0"
django-rest-swagger = "^2.2.0"
PyYAML = "^5.4.1"
uritemplate = "^3.0.1"
celery = "^5.0.5"
psycopg2-binary = "^2.8.6"
dj-database-url = "^0.5.0"
matplotlib = "^3.4.1"
statsmodels = "^0.12.2"
dill = "^0.3.3"
sentry-sdk = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
